requirements.txt
aiofiles
asyncpg
fastapi
httpx
pyyaml
SQLAlchemy
uvicorn